backend/requirements.txt
fastapi 
uvicorn 
motor 
passlib 
bcrypt 
python-multipart 
python-jose[cryptography]
